Store count of frames to pop in FRAME_POP flags
This both simplifies the tracking API, pushing some encoding complexity
down into the record writer, and also saves space in the output file.

+++ src/memray/_memray/records.h
@@ -197,7 +197,7 @@ struct FramePush
 
 struct FramePop
 {
-    uint8_t count;
+    size_t count;
 };
 
 struct UnresolvedNativeFrame
